+++ tests/Rules/ValidIbanTest.php
@@ -2,7 +2,7 @@
 
 namespace Milwad\LaravelValidate\Tests\Rules;
 
-use Milwad\LaravelValidate\Rules\Country;
+use Milwad\LaravelValidate\Utils\Country;
 use Milwad\LaravelValidate\Rules\ValidIban;
 use Milwad\LaravelValidate\Tests\BaseTest;
